@@ -1,68 +1,61 @@
 import FeedbackOptions from './FeedBack/FeedbackOptions/FeedbackOptions';
-import React, { Component } from 'react';
+import { useState } from 'react';
 import Statistics from './FeedBack/statistics/Statistics';
 import Section from './FeedBack/section/Section';
 import Notification from './FeedBack/notification/Notification';
 
-export class App extends Component {
-  state = {
-    good: 0,
-    neutral: 0,
-    bad: 0,
+export const App = () => {
+  const [good, setGood] = useState(0);
+  const [neutral, setNeutral] = useState(0);
+  const [bad, setBad] = useState(0);
+
+  const handleFeedback = feetback => {
+    switch (feetback) {
+      case 'good':
+        setGood(prevstate => prevstate + 1);
+        break;
+      case 'neutral':
+        setNeutral(prevstate => prevstate + 1);
+        break;
+      case 'bad':
+        setBad(prevstate => prevstate + 1);
+        break;
+      default:
+        console.log('error');
+    }
   };
 
-  handleFeedback = e => {
-    const eventName = e.target.textContent.toLowerCase();
-    this.setState(prevState => {
-      return {
-        [eventName]: prevState[eventName] + 1,
-      };
-    });
+  const countTotalFeedback = () => {
+    return good + neutral + bad;
   };
 
-  countTotalFeedback = () => {
-    const ArrayValuses = Object.values(this.state);
-
-    const sumWithInitial = ArrayValuses.reduce(
-      (previousValue, currentValue) => previousValue + currentValue
-    );
-
-    return sumWithInitial;
+  const countPositiveFeedbackPercentage = () => {
+    return good ? Math.round((good / countTotalFeedback()) * 100) : 0;
   };
 
-  countPositiveFeedbackPercentage = () => {
-    return this.state.good
-      ? Math.round((this.state.good / this.countTotalFeedback()) * 100)
-      : 0;
-  };
+  const ArrayState = ['good', 'neutral', 'bad'];
 
-  render() {
-    const ArrayState = Object.keys(this.state);
-
-    const { good, neutral, bad } = this.state;
-
-    return (
-      <div>
-        <Section title={'Please leave feedback'}>
-          <FeedbackOptions
-            options={ArrayState}
-            onLeaveFeedback={this.handleFeedback}
+  return (
+    <div>
+      <Section title={'Please leave feedback'}>
+        <FeedbackOptions
+          options={ArrayState}
+          onLeaveFeedback={handleFeedback}
+        />
+      </Section>
+      <Section title={'Section'}>
+        {good || neutral || bad ? (
+          <Statistics
+            good={good}
+            neutral={neutral}
+            bad={bad}
+            total={countTotalFeedback}
+            positivePercentage={countPositiveFeedbackPercentage}
           />
-        </Section>
-        <Section title={'Section'}>
-          {good || neutral || bad ? (
-            <Statistics
-              good={good}
-              neutral={neutral}
-              bad={bad}
-              total={this.countTotalFeedback}
-              positivePercentage={this.countPositiveFeedbackPercentage}
-            />
-          ) : (
-            <Notification massage={'There is no feedback'} />
-          )}
-        </Section>
-      </div>
-    );
-  }
-}
+        ) : (
+          <Notification massage={'There is no feedback'} />
+        )}
+      </Section>
+    </div>
+  );
+};
