@@ -3,12 +3,16 @@ import { FBOContainer } from './FeedBackOptions.styled';
 import propTypes from 'prop-types';
 
 const FeedbackOptions = ({ options, onLeaveFeedback }) => {
-  console.log(options);
   return (
     <FBOContainer>
       {options.map(item => {
         return (
-          <button key={item} onClick={onLeaveFeedback}>
+          <button
+            key={item}
+            onClick={() => {
+              onLeaveFeedback(item);
+            }}
+          >
             {item}
           </button>
         );
@@ -20,6 +24,6 @@ const FeedbackOptions = ({ options, onLeaveFeedback }) => {
 export default FeedbackOptions;
 
 FeedbackOptions.propTypes = {
-  option: propTypes.arrayOf(propTypes.string).isRequired,
+  options: propTypes.arrayOf(propTypes.string).isRequired,
   onLeaveFeedback: propTypes.func.isRequired,
 };
